library/017-AI/近屿/aigc_learning_path/aigc_tm_level/A2/A2.1大模型提示工程/requirements.txt
python-dotenv==1.0.0
matplotlib==3.8.3
promptbench==0.0.2
dspy-ai==2.4.10
openai
httpx==0.27.2